requirements.txt
scholarly==1.2.2
pandas==1.3.1
habanero==0.7.4
pickle5==0.0.11
bibtexparser==1.2.0
python-Levenshtein==0.12.2
numpy==1.21.1
